requirements.txt
newspaper3k>=0.2.8
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml_html_clean>=0.4.1
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.0
notion-client>=2.2.1
pyperclip>=1.8.0
feedparser>=6.0.10
